Rename call.h to call_op_set.h and call_wrapper.h to call.h
