+++ build.yaml
@@ -1222,7 +1222,7 @@ filegroups:
   - include/grpcpp/impl/codegen/byte_buffer.h
   - include/grpcpp/impl/codegen/call.h
   - include/grpcpp/impl/codegen/call_hook.h
-  - include/grpcpp/impl/codegen/call_wrapper.h
+  - include/grpcpp/impl/codegen/call_op_set.h
   - include/grpcpp/impl/codegen/callback_common.h
   - include/grpcpp/impl/codegen/channel_interface.h
   - include/grpcpp/impl/codegen/client_callback.h
@@ -1259,7 +1259,6 @@ filegroups:
 - name: grpc++_codegen_base_src
   language: c++
   src:
-  - src/cpp/codegen/call_wrapper.cc
   - src/cpp/codegen/codegen_init.cc
   uses:
   - grpc++_codegen_base
