+++ gRPC-C++.podspec
@@ -128,7 +128,7 @@ Pod::Spec.new do |s|
                       'include/grpcpp/impl/codegen/byte_buffer.h',
                       'include/grpcpp/impl/codegen/call.h',
                       'include/grpcpp/impl/codegen/call_hook.h',
-                      'include/grpcpp/impl/codegen/call_wrapper.h',
+                      'include/grpcpp/impl/codegen/call_op_set.h',
                       'include/grpcpp/impl/codegen/callback_common.h',
                       'include/grpcpp/impl/codegen/channel_interface.h',
                       'include/grpcpp/impl/codegen/client_callback.h',
@@ -218,7 +218,6 @@ Pod::Spec.new do |s|
                       'src/cpp/util/status.cc',
                       'src/cpp/util/string_ref.cc',
                       'src/cpp/util/time_cc.cc',
-                      'src/cpp/codegen/call_wrapper.cc',
                       'src/cpp/codegen/codegen_init.cc',
                       'src/core/lib/gpr/alloc.h',
                       'src/core/lib/gpr/arena.h',
