+++ grpc.gyp
@@ -1420,7 +1420,6 @@
         'third_party/nanopb/pb_common.c',
         'third_party/nanopb/pb_decode.c',
         'third_party/nanopb/pb_encode.c',
-        'src/cpp/codegen/call_wrapper.cc',
         'src/cpp/codegen/codegen_init.cc',
       ],
     },
@@ -1501,7 +1500,6 @@
         'test/cpp/util/string_ref_helper.cc',
         'test/cpp/util/subprocess.cc',
         'test/cpp/util/test_credentials_provider.cc',
-        'src/cpp/codegen/call_wrapper.cc',
         'src/cpp/codegen/codegen_init.cc',
       ],
     },
@@ -1522,7 +1520,6 @@
         'test/cpp/util/byte_buffer_proto_helper.cc',
         'test/cpp/util/string_ref_helper.cc',
         'test/cpp/util/subprocess.cc',
-        'src/cpp/codegen/call_wrapper.cc',
         'src/cpp/codegen/codegen_init.cc',
       ],
     },
@@ -1573,7 +1570,6 @@
         'third_party/nanopb/pb_common.c',
         'third_party/nanopb/pb_decode.c',
         'third_party/nanopb/pb_encode.c',
-        'src/cpp/codegen/call_wrapper.cc',
         'src/cpp/codegen/codegen_init.cc',
       ],
     },
