+++ tools/doxygen/Doxyfile.c++
@@ -943,7 +943,7 @@ include/grpcpp/impl/codegen/async_unary_call.h \
 include/grpcpp/impl/codegen/byte_buffer.h \
 include/grpcpp/impl/codegen/call.h \
 include/grpcpp/impl/codegen/call_hook.h \
-include/grpcpp/impl/codegen/call_wrapper.h \
+include/grpcpp/impl/codegen/call_op_set.h \
 include/grpcpp/impl/codegen/callback_common.h \
 include/grpcpp/impl/codegen/channel_interface.h \
 include/grpcpp/impl/codegen/client_callback.h \
